relocating attribute definitions from type to common since geometry collection will need it.
