+++ engine/runtime/action-processor/src/feature/reader/citygml/reader.rs
@@ -23,8 +23,10 @@ use reearth_flow_runtime::{
     node::DEFAULT_PORT,
 };
 use reearth_flow_types::{
-    conversion::nusamai::entity_to_geometry, geometry::Geometry, lod::LodMask, Attribute,
-    AttributeValue, CitygmlFeatureExt, CompiledCode, Feature,
+    conversion::nusamai::{entity_to_geometry, from_nusamai_citygml_value},
+    geometry::Geometry,
+    lod::LodMask,
+    Attribute, AttributeValue, CitygmlFeatureExt, CompiledCode, Feature,
 };
 use url::Url;
 
@@ -201,8 +203,7 @@ fn emit_flat_entity(
         }
     }
 
-    let citygml_attributes =
-        AttributeValue::Map(AttributeValue::from_nusamai_citygml_value(&ent.root));
+    let citygml_attributes = AttributeValue::Map(from_nusamai_citygml_value(&ent.root));
     let geometry: Geometry = entity_to_geometry(ent, root_needs_reconstruction)
         .map_err(|e| FeatureProcessorError::FileCityGmlReader(format!("{e:?}")))?;
     let mut feature: Feature = geometry.into();
