+++ engine/runtime/action-sink/src/file/gltf.rs
@@ -14,7 +14,7 @@ use reearth_flow_runtime::event::EventHub;
 use reearth_flow_runtime::executor_operation::{ExecutorContext, NodeContext};
 use reearth_flow_runtime::node::{Port, Sink, SinkFactory, DEFAULT_PORT};
 use reearth_flow_types::material::{self, Material};
-use reearth_flow_types::{Code, GeometryType};
+use reearth_flow_types::{attribute_value_to_citygml_attribute, Code, GeometryType};
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
@@ -122,7 +122,10 @@ impl TryFrom<&ClassFeatures> for Schema {
             .iter()
             .filter(|(_, v)| v.convertible_nusamai_type_ref())
         {
-            attributes.insert(k.to_string(), v.clone().into());
+            attributes.insert(
+                k.to_string(),
+                attribute_value_to_citygml_attribute(v.clone()),
+            );
         }
         schema.types.insert(
             feature_type,
