+++ engine/runtime/common/Cargo.toml
@@ -29,11 +29,14 @@ fastxml.workspace = true
 futures.workspace = true
 home.workspace = true
 image.workspace = true
+indexmap.workspace = true
 jsonpath_lib.workspace = true
+nutype.workspace = true
 once_cell.workspace = true
 parking_lot.workspace = true
 rayon.workspace = true
 regex.workspace = true
+schemars.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 serde_yaml.workspace = true
