+++ engine/runtime/types/src/conversion.rs
@@ -1,2 +1,3 @@
 pub mod geojson;
 pub mod nusamai;
+pub mod rhai;
