+++ engine/runtime/types/src/lib.rs
@@ -13,6 +13,11 @@ pub mod material;
 pub mod workflow;
 
 pub use attribute::*;
+pub use conversion::nusamai::{
+    attribute_value_to_citygml_attribute, attribute_value_to_citygml_type_ref,
+    from_nusamai_citygml_value,
+};
+pub use conversion::rhai::attribute_value_from_rhai;
 pub use expr::*;
 pub use feature::*;
 pub use file::*;
